% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_removing2.R
\name{all_items_from_var_list2}
\alias{all_items_from_var_list2}
\title{Internal: collect all items appearing in var_list measurement models}
\usage{
all_items_from_var_list2(var_list, observed)
}
\description{
Internal: collect all items appearing in var_list measurement models
}
\keyword{internal}
