% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_removing.R
\name{alpha_from_cov}
\alias{alpha_from_cov}
\title{Internal: Cronbach's alpha from a covariance matrix}
\usage{
alpha_from_cov(S)
}
\description{
Internal: Cronbach's alpha from a covariance matrix
}
\keyword{internal}
