% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_removing2.R
\name{alpha_from_cov2}
\alias{alpha_from_cov2}
\title{Internal: Cronbach's alpha from a covariance matrix}
\usage{
alpha_from_cov2(S)
}
\description{
Internal: Cronbach's alpha from a covariance matrix
}
\keyword{internal}
