% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_table.R
\name{copy_table}
\alias{copy_table}
\title{Copy the Last Printed Data Frame to Clipboard}
\usage{
copy_table(row.names = FALSE, col.names = TRUE, sep = "\\t", ...)
}
\arguments{
\item{row.names}{Logical. Should row names be included? Default is FALSE.}

\item{col.names}{Logical. Should column names be included? Default is TRUE.}

\item{sep}{Field separator string. Default is tab character.}

\item{...}{Additional arguments passed to \code{\link{write.table}}.}
}
\value{
Invisibly returns the copied object.
}
\description{
Copies the most recently evaluated data frame or matrix
to the system clipboard in tab-separated format, suitable
for pasting into spreadsheet applications.
}
\examples{
\dontrun{
head(mtcars)
copy_table()
}

}
