% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pot_data.R
\name{extract_plot_data}
\alias{extract_plot_data}
\title{Exact plot data}
\usage{
extract_plot_data(model, model_label)
}
\arguments{
\item{model}{object returned by \code{readModels} for mixture model}

\item{model_label}{index of the latent class model, e.g., 1-class model, 2-class model}
}
\value{
a data.frame
}
\description{
Exact plot data which combines both means and probabilities
}
\examples{


# Applying the function to all model objects within the list
plot_data_pre <- map2(res_pre, 1:5, extract_plot_data) \%>\%
  bind_rows() \%>\%
  filter(!param == "DAYS_GAP") \%>\%
  mutate(param = recode(param,
                        "TEC_RECOG" = "Recognition",
                        "TEC_EXT" = "External causes",
                        "TEC_DES" = "Desire",
                        "TEC_BEL" = "Belief",
                        "TEC_REM" = "Reminder",
                        "TEC_REG" = "Regulation",
                        "TEC_HID" = "Hiding"),
         param = factor(param, levels = c("Recognition", "External causes", "Desire", "Belief", "Reminder", "Regulation", "Hiding"))) # Change the order of variables for plotting
}
