% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggLin.R
\name{ggLin}
\alias{ggLin}
\alias{theme_lin}
\title{Jennifer's ggplot2 Theme}
\usage{
theme_lin(...)
}
\arguments{
\item{...}{Passed to [ggplot2::theme()]}
}
\description{
I find that I am copying and psting the same long command for
my favorite theme settings, so I figure I'd write that into this package
to simplify the process.
}
\details{
The specifications of this theme is as follows: \cr
theme_classic()+   \cr
theme(   \cr
plot.title         = element_text(hjust = 0.5, size = 24, colour="black"),  \cr
 plot.subtitle      = element_text(hjust = 0.5, size = 18, colour="black"),  \cr
 legend.title       = element_text(hjust = 0.5, size = 16, colour="black"),  \cr
 plot.caption       = element_text(size = 12, colour="black"),  \cr
 axis.title         = element_text(size = 16, colour="black"),  \cr
 axis.text.x        = element_text(size = 14, colour="black"),  \cr
 axis.text.y        = element_text(size = 14, colour="black"),  \cr
 legend.title.align = 0.5)
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(y=mpg, x=disp, color=cyl)) +
  geom_point() +
  theme_lin()


}
