% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_color.r
\name{lin_color}
\alias{lin_color}
\title{Custom colors}
\usage{
lin_color(...)
}
\arguments{
\item{...}{The color name(s) to use. If left blank, it will return the entire my_color vector}
}
\description{
Custom colors
}
\examples{
plot(rnorm(100), col = lin_color("yanzhichong"))

}
