% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Operators.R
\name{Operators}
\alias{Operators}
\alias{\%nin\%}
\alias{\%NIN\%}
\alias{\%IN\%}
\title{Operators}
\usage{
x \%nin\% y

x \%NIN\% y

x \%IN\% y
}
\arguments{
\item{x}{a vector}

\item{y}{a vector to match}
}
\value{
logical vecotor of items in x not in y

logical vecotor of items in x not in y, omits NAs

logical vecotor of items in x in y, omits NAs
}
\description{
Some useful operators outside of the standard ones in R.
}
\examples{
y = c(3, 4, 5, NA)

# Not In -- and omits NA
y \%nin\% 3    # FALSE  TRUE  TRUE  TRUE
y \%NIN\% 3    # FALSE  TRUE  TRUE    NA

# IN -- Omits NA
y \%in\% 3     # TRUE FALSE FALSE FALSE
y \%IN\% 3     # TRUE FALSE FALSE    NA

}
