% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan.r
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Custom colors}
\usage{
plot_manhattan(tbl, var_p, var_x)
}
\arguments{
\item{tbl}{The data.frame or tibble used to plot}

\item{var_p}{The column containing p-value}

\item{var_x}{The column used for x-axis}
}
\description{
Custom colors
}
\examples{
plot_manhattan(data.frame(a = sample(seq(0, 1, 0.1), 10), b = 1:10), a, b)

}
