% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PNG.R
\name{plot_PNG}
\alias{plot_PNG}
\title{Save ggplot object as a PNG file}
\usage{
plot_PNG(p, file, width = 3000, height = 3000, res = 300)
}
\arguments{
\item{p}{the ggplot plot object}

\item{file}{a string, the name/path of the PNG file to save}

\item{width}{width in px of the output figure}

\item{height}{height in px of the output figure}

\item{res}{resolution of the output}
}
\description{
Save ggplot object as a PNG file
}
\examples{
library(ggplot2)
p <- ggplot(mapping = aes(x = rand(100), y = rand(100)))
plot_PNG(p, "sample.PNG")

}
