% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReverseCode.R
\name{ReverseCode}
\alias{ReverseCode}
\alias{reverse_code}
\title{Reverse Coding Variables}
\source{
\url{https://github.com/jamesmartherus/martherus}
}
\usage{
reverse_code(var)
}
\description{
I never actually know how to do them, so I google this every time.
Perhaps its time to settle this once and for all.
}
\details{
Thanks goes to James Martherus.
}
\examples{
x1 <- c(1, 2, 3, 4, NA, 5)
reverse_code(x1)   #c(5, 4, 3, 2, NA, 1)

x2 <- c(0, 1, 2, NA, 4, 7)
recode_code(x2)    #c(7, 6, 5, NA, 3, 0)

}
