% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_with_proportions.R
\name{summary_with_proportions}
\alias{summary_with_proportions}
\title{Summary with proportions
Title Fit indices and proportion of each class in LCA analysis}
\usage{
summary_with_proportions(res)
}
\arguments{
\item{res}{object returned by \code{readModels}.}
}
\value{
a data.frame
}
\description{
Summary with proportions
Title Fit indices and proportion of each class in LCA analysis
}
\examples{
summary_with_proportions(res_pre) \%>\%
      mutate(Title = str_replace(Title, " model;", "")) \%>\%
      kable(digits = 2, caption = "Table x. Model fit summary.") \%>\%
      kable_classic(html_font = "Cambria")
}
