% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstat.R
\name{SummaryStats}
\alias{SummaryStats}
\alias{modeNA}
\alias{meanNA}
\alias{wMeanNA}
\title{Summary Statistics Calculations}
\source{
\url{https://github.com/jbullock35/Bullock}
}
\usage{
modeNA(x)

meanNA(x)

wMeanNA(x, w)
}
\arguments{
\item{x}{a vector}

\item{w}{a weight variable}
}
\description{
Common functions for calculating central tendancies but with
NA parameters set to TRUE unlike the defaults.
}
\details{
Credits to John Bullock for the inspiration. Some of this is from his
Bullock package, but others are my own.
}
\examples{
x <- c(1, 1, 2, 3, 5, 8, 13, 21, NA, NA, NA)
w <- c(0, 0, 0, 1, 1, 2, 2, 2, 0, 0, 0)

# Mode
modeNA(x)    # 1

# Mean and Weighted Mean
meanNA(x)    # 6.75
wMeanNA(x)   # 11.5


}
