% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscoreCI.R
\name{tscoreCI}
\alias{tscoreCI}
\alias{tCIupper}
\alias{tCIlower}
\alias{tCI}
\alias{tMOE}
\title{One Sample Confidence Interval from a t Distribution}
\usage{
tCIupper(x, se, ci, n)

tCIlower(x, se, ci, n)

tCI(x, se, ci, n)

tMOE(se, ci, n)
}
\arguments{
\item{x}{x-bar for observed mean}

\item{se}{Standard Error -- see se() from zscoreCI section}

\item{ci}{Confidence Interval Level}

\item{n}{Number of participants}
}
\description{
Helpful functions to calculate a one-sample confidence interval
using a t-distribution
}
\details{
NOTE: This is not NOT intended from a z-distribution.
See relevant zscoreCI file for those calculations
}
\examples{
x = 3.5
sd = 2
n = 25

se <- se(sd, n)  # 0.4

ci = 0.95  # for 95\% Confidence Interval

tCIupper(x, se, ci, n) # 4.325559
tCIlower(x, se, ci, n) # 2.674441
tMOE(se, ci, n)        # 0.8255594


}
\seealso{
zCI(), zCIupper(), zCIlower()
}
