% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tstats.R
\name{Tstats}
\alias{Tstats}
\alias{t}
\alias{t.welch}
\title{Two Sample T Test and Confidence Interval}
\usage{
t(x1, s1, n1, x2, s2, n2)

t.welch(x1, s1, n1, x2, s2, n2)
}
\arguments{
\item{x1}{mean for sample 1}

\item{s1}{sd for sample 1}

\item{n1}{number on sample 1}

\item{x2}{mean for sample 2}

\item{s2}{sd for sample 2}

\item{n2}{number on sample 2}
}
\description{
For PS 403 purposes. In real life, you should always just use
\code{t.test()}
}
\examples{
x1 = 5
x2 = 8
s1 = 0.6
s2 = 0.4
n1 = 30
n2 = 25

t(x1, s1, n1, x2, s2, n2)

}
