% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pridict_sim.R
\name{validate_simulation}
\alias{validate_simulation}
\title{Validate Simulated Data}
\usage{
validate_simulation(df, outcome_col = "outcome")
}
\arguments{
\item{df}{A data frame containing the outcome and predictors.}

\item{outcome_col}{A string specifying the name of the outcome column. Defaults to "outcome".}
}
\value{
A list containing the linear model object and the data frame.
}
\description{
Runs a linear regression model and checks frequency distributions to validate
the generated predictor variables against the outcome.
}
