% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize.R
\name{winsorize}
\alias{winsorize}
\title{Helper function to recode outliers}
\usage{
winsorize(x)
}
\arguments{
\item{x}{column of a data.frame}
}
\value{
a new columns after mutate function
}
\description{
Winsorizing value which outside the 3 SD boundary.
}
\examples{
data_clean <- data \%>\%
mutate(
# Recode rt and acc value
      rt_recode = winsorize(rt_m),
      acc_recode = winsorize(acc_m))



}
