% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zprop.R
\name{Zprop}
\alias{Zprop}
\alias{prop.se}
\alias{p.hat}
\alias{z.prop}
\title{Two Sample Proportions Test and Confidence Interval}
\usage{
prop.se(p, n)

p.hat(p1, n1, p2, n2)

z.prop(p1, n1, p2, n2)
}
\arguments{
\item{p}{pi for a proportion, usually denoted as pi_0}

\item{n}{number of responses}

\item{p1}{probability for sample 1}

\item{n1}{number on sample 1}

\item{p2}{probability for sample 2}

\item{n2}{number on sample 2}
}
\description{
For use with proportions and z-scores.
}
\details{
When working with a one-sample proportion, it is suffucient to use
\code{prop.se()} from here and the \code{zscore()} function, replacing
the notions of \code{x} and \code{means} with the proper values for proportions

When working with these functions, it is CRUCIAL that you keep in mind
what vales are for what. I would recommend setting up code similiar to the
example to keep values in order.
}
\examples{
pi = 0.5
n = 100

p1 = 0.25
n1 = 200
p2 = 0.45
n2 = 350

prop.se(pi, n)
p.hat(p1, n1, p2, n2)


}
