% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{z-score Calculations}
\usage{
zscore(x, mean, sd)
}
\arguments{
\item{x}{the observation}

\item{mean}{mean of interest -- can be sample or ppulation depending on zscore interest}

\item{sd}{standard deviation or standard error, depending on context}
}
\description{
Calculating a standard score in Base R can be hard.
}
\examples{
zscore(10, 15, 2)

}
