% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscoreCI.R
\name{zscoreCI}
\alias{zscoreCI}
\alias{se}
\alias{zCIupper}
\alias{zCIlower}
\alias{zCI}
\alias{zMOE}
\title{One Sample Confidence Interval from a Normal Distribution}
\usage{
se(sd, n)

zCIupper(x, se, ci)

zCIlower(x, se, ci)

zCI(x, se, ci)

zMOE(se, ci)
}
\arguments{
\item{sd}{Standard Deviation -- reference sdNA()}

\item{n}{Number of participants}

\item{x}{x-bar for observed mean}

\item{se}{Standard Error -- see se()}

\item{ci}{Confidence Interval Level}
}
\description{
Helpful functions to calculate a one-sample confidence interval
using a z-distribution
}
\details{
NOTE: This is not NOT intended from a t-distribution.
See relavant tscoreCI file for those calculations
}
\examples{
x = 3.5
sd = 2
n = 25

se <- se(sd, n)  # 0.4

ci = 0.95  # for 95\% Confidence Interval

zCIupper(x, se, ci) # 4.283986
zCIlower(x, se, ci) # 2.716014
zMOE(se, ci)        # 0.7839856

}
\seealso{
sdNA()
}
