% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa.R
\name{apa}
\alias{apa}
\title{APA a table}
\usage{
apa(x, title = NULL, notes = NULL, digits = 3)
}
\arguments{
\item{x}{data.frame}

\item{title}{title of table}

\item{notes}{footnote of table}

\item{digits}{digits of the value within the table}
}
\value{
a data.frame
}
\description{
Using kable and kable_classic to create a apa format table
}
\examples{
# Format the first few rows of mtcars as an APA style table
apa(head(mtcars, 5), title = "Descriptive Statistics for mtcars Sample")
}
