% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cfa_model.R
\name{create_cfa_model}
\alias{create_cfa_model}
\title{Create lavaan model}
\usage{
create_cfa_model(lav_name, pattern, var_index)
}
\arguments{
\item{lav_name}{names of latent variable}

\item{pattern}{pattern of the model structure}

\item{var_index}{which item should be counted into the model}
}
\value{
a string contains the model structure
}
\description{
A more easy way to create lavaan style cfa model syntax
}
\examples{
# 1) Single-factor example:
model1 <- create_cfa_model("x", "x{i}", 1:3)
# which would return model1 <- 'x =~ x1 + x2 + x3`

# Evaluate fit and reliability:
lavaan::cfa(model1, data = data)
reliability_table(model1, data = mydata, digits = 3)

# 2) Multi-factor model, then reliability:
model2 <- paste(
  create_cfa_model("x", "x{i}", 1:5),
  create_cfa_model("m", "m{i}", 1:3),
  create_cfa_model("z", "z{i}", 1:4),
  collapse = "\n"
)
# Now get reliability for the combined model:
reliability_table(model2, data = mydata, digits = 3)


}
