% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_table.R
\name{fit_table}
\alias{fit_table}
\title{Model fit table}
\usage{
fit_table(
  modelList,
  indices = c("Filename", "ChiSqM_Value", "ChiSqM_DF", "CFI", "TLI", "RMSEA_Estimate",
    "SRMR"),
  ref = c("last", "first"),
  diffTest = TRUE,
  digits = 3
)
}
\arguments{
\item{modelList}{Generated by readModels in MplusAutomation, can be a single object or a list}

\item{indices}{which indices be reported}

\item{ref}{Reference group in the model comparison, default is the latest model}

\item{diffTest}{Should the comparison included? Including change of CFI, TLI, and Chi-square.}

\item{digits}{the digits of each indice, degree of freedom is always the integer.}
}
\value{
a \code{data.table} object
}
\description{
Use the information from \code{MplusAutomation} to create a model fit table
}
\examples{

fit_table(model2),
    keepCols = bruceR::cc("Title, Observations, ChiSqM_Value, ChiSqM_DF, CFI, TLI, RMSEA_Estimate, SRMR")) \%>\%
    mutate(Title = "Coefficients", digits = 3)

fit_table(list(models$model1, models$model2, models$model2)),
    keepCols = bruceR::cc("Title, Observations, ChiSqM_Value, ChiSqM_DF, CFI, TLI, RMSEA_Estimate, SRMR")) \%>\%
    mutate(Title = bruceR::cccc("model1, model2, model3"), digits = 3)
}
